refactor: complete migration from _redact_compat to SDK redact module
Complete the migration started in the initial commit:
- Update all remaining imports from openhands.utils._redact_compat to
  openhands.sdk.utils.redact (sanitize_config, redact_text_secrets,
  redact_url_params)
- Delete the _redact_compat.py shim file (no longer needed since SDK
  v1.16.1+ has all required functions)
- Rename test file from test_redact_compat.py to test_redact.py and
  update imports to test SDK functions directly
- Adjust one test case to use a longer key matching SDK's 20+ char
  regex requirement for sk-oh- tokens

Files migrated:
- openhands/app_server/event_callback/set_title_callback_processor.py
- openhands/app_server/event_callback/event_callback_models.py
- openhands/app_server/app_conversation/live_status_app_conversation_service.py
- openhands/runtime/action_execution_server.py
- openhands/mcp/utils.py
