+++ openhands/app_server/event_callback/event_callback_models.py
@@ -16,14 +16,13 @@ from openhands.app_server.event_callback.event_callback_result_models import (
     EventCallbackResultStatus,
 )
 
-# TODO(OpenHands/evaluation#418): import from openhands.sdk.utils.redact
-from openhands.app_server.utils._redact_compat import redact_text_secrets
 from openhands.sdk import Event
 from openhands.sdk.utils.models import (
     DiscriminatedUnionMixin,
     OpenHandsModel,
     get_known_concrete_subclasses,
 )
+from openhands.sdk.utils.redact import redact_text_secrets
 
 _logger = logging.getLogger(__name__)
 if TYPE_CHECKING:
