+++ openhands/app_server/event_callback/set_title_callback_processor.py
@@ -19,12 +19,11 @@ from openhands.app_server.event_callback.event_callback_result_models import (
 from openhands.app_server.services.injector import InjectorState
 from openhands.app_server.user.specifiy_user_context import ADMIN, USER_CONTEXT_ATTR
 
-# TODO(OpenHands/evaluation#418): import from openhands.sdk.utils.redact
-from openhands.app_server.utils._redact_compat import redact_text_secrets
 from openhands.app_server.utils.docker_utils import (
     replace_localhost_hostname_for_docker,
 )
 from openhands.sdk import Event, MessageEvent
+from openhands.sdk.utils.redact import redact_text_secrets
 
 _logger = logging.getLogger(__name__)
